# smart-master/smart-web/swagger.json
{"swagger":"2.0","info":{"description":"smart接口介绍标题","version":"1.0","title":"smart接口介绍标题"},"host":"localhost","basePath":"/","tags":[{"name":"menu-controller","description":"Menu Controller"},{"name":"role-controller","description":"Role Controller"},{"name":"user-controller","description":"用户基本操作"}],"paths":{"/menus/getMenus":{"post":{"tags":["menu-controller"],"summary":"获取自己的资源信息","operationId":"getMenusUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"username","description":"用户名","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/MplMenu"}}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/role/addRole":{"post":{"tags":["role-controller"],"summary":"添加用户角色","operationId":"addRoleUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"roleName","in":"query","description":"roleName","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Msg"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/role/deleteRole":{"post":{"tags":["role-controller"],"summary":"根据id删除角色数据","operationId":"deleteRoleUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"roleId","in":"query","description":"roleId","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Msg"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/role/getRole":{"post":{"tags":["role-controller"],"summary":"根据token获取角色信息","operationId":"getRoleUsingPOST","consumes":["application/json"],"produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Msg"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/role/getRoles":{"post":{"tags":["role-controller"],"summary":"获取全部角色信息","operationId":"getRolesUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"pageNum","in":"query","description":"页码 缺省1","required":false,"type":"integer","format":"int32"},{"name":"pageSize","in":"query","description":"每页显示条数 缺省10","required":false,"type":"integer","format":"int32"},{"name":"roleName","in":"query","description":"角色名称","required":false,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/CommonResult"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/role/setUserRole":{"post":{"tags":["role-controller"],"summary":"绑定角色和用户","operationId":"setUserRoleUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"userid","in":"query","description":"用户id","required":false,"type":"string"},{"name":"roleid","in":"query","description":"角色id","required":false,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Msg"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/role/updateRole":{"post":{"tags":["role-controller"],"summary":"修改角色信息","operationId":"updateRoleUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"id","in":"query","description":"角色id","required":false,"type":"integer","format":"int64"},{"name":"roleName","in":"query","description":"角色名称","required":false,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Msg"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/user/info":{"get":{"tags":["user-controller"],"summary":"获取用户信息","operationId":"infoUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Msg"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/user/login":{"post":{"tags":["user-controller"],"summary":"登陆接口","operationId":"loginUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"vo","description":"vo","required":true,"schema":{"$ref":"#/definitions/登陆"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Msg"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/user/logout":{"post":{"tags":["user-controller"],"summary":"登出接口","operationId":"logoutUsingPOST","consumes":["application/json"],"produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Msg"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/user/register":{"post":{"tags":["user-controller"],"summary":"注册接口","operationId":"registerUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"user","description":"user","required":true,"schema":{"$ref":"#/definitions/注册"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Msg"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/user/resetPassword":{"post":{"tags":["user-controller"],"summary":"重置用户密码","operationId":"resetPasswordUsingPOST","consumes":["application/json"],"produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Msg"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/user/resource":{"post":{"tags":["user-controller"],"summary":"获取用户资源","operationId":"getResourceUsingPOST","consumes":["application/json"],"produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Msg"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/user/updatePassword":{"post":{"tags":["user-controller"],"summary":"修改密码","operationId":"updatePasswordUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"vo","description":"vo","required":true,"schema":{"$ref":"#/definitions/修改密码"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Msg"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}}},"definitions":{"CommonResult":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"data":{"type":"array","items":{"type":"object"}},"msg":{"type":"string"},"pageNum":{"type":"integer","format":"int32"},"pageSize":{"type":"integer","format":"int32"},"timestamp":{"type":"integer","format":"int64"},"total":{"type":"integer","format":"int32"},"totalPage":{"type":"integer","format":"int64"}},"title":"CommonResult"},"MplMenu":{"type":"object","properties":{"childMenus":{"type":"array","items":{"$ref":"#/definitions/MplMenu"}},"createTime":{"type":"string","format":"date-time"},"createUser":{"type":"string"},"icon":{"type":"string"},"id":{"type":"integer","format":"int64"},"isShow":{"type":"string"},"menuName":{"type":"string"},"parentId":{"type":"integer","format":"int64"},"ramark":{"type":"string"},"sequence":{"type":"string"},"type":{"type":"string"},"updateTime":{"type":"string","format":"date-time"},"updateUser":{"type":"string"},"url":{"type":"string"}},"title":"MplMenu"},"Msg":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"data":{"type":"object"},"msg":{"type":"string"},"timestamp":{"type":"integer","format":"int64"}},"title":"Msg"},"修改密码":{"type":"object","properties":{"newpassword":{"type":"string","description":"新密码","readOnly":true,"allowEmptyValue":false},"password":{"type":"string","description":"原密码","readOnly":true,"allowEmptyValue":false},"username":{"type":"string","description":"用户名","readOnly":true,"allowEmptyValue":false}},"title":"修改密码"},"注册":{"type":"object","properties":{"password":{"type":"string"},"username":{"type":"string"}},"title":"注册"},"登陆":{"type":"object","required":["password","username"],"properties":{"password":{"type":"string","description":"密码","allowEmptyValue":false},"username":{"type":"string","description":"用户名","allowEmptyValue":false}},"title":"登陆"}}}